Refactor of Phase Platform
The original version of the script is still here under the name
OldPhasePlatform.

The new method handles detection of events in a parent class called
BeatObject, so only the effects of the class need to be in each specific
file now. More variables are also exposed so that they can be tweaked
from the editor (what types of beat will trigger onBeat( ).)

Things to remember for these other scripts:
- void Update( ) must contain a call to base.Update( )
- All effects that occur on a particular beat must be within "override
void onBeat( )"

KNOWN ISSUES (That I hope to get some help on at the meeting tonight):
- It would appear that both the old and new versions of the code will
only operate on one object in the scene at the moment.