Began Work Fixing Beat Signal System
So, apparently C# events can only be caught by one object. I've started
making a new script called BeatSignalManager that will catch these
signals, and call a certain method on every BeatObject in the scene.

Another commit will be added tomorrow that will finish this script and
make it functional. (Need to create an array of references to every
BeatObject in the scene).